% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_stan.R
\name{lm_stan}
\alias{lm_stan}
\title{Bayesian linear regression with Stan}
\usage{
lm_stan(x, y, ...)
}
\arguments{
\item{x}{Numeric matrix of input values. This matrix should not contain
a column of 1's.}

\item{y}{Numeric vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian linear regression with Stan
}
\examples{
\dontrun{
x <- runif(30, 0, 10)
y <- 1 + 2*x + rnorm(30)
lmFit <- lm_stan(x, y)
}

}
