% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlmfit-class.R
\name{print.rlmfit}
\alias{print.rlmfit}
\title{Print method for rlmfit objects}
\usage{
\method{print}{rlmfit}(object, pars, probs = c(0.025, 0.25, 0.5, 0.75,
  0.975), digits_summary = 2, include = TRUE, ...)
}
\description{
Print a summary of parameter estimates and MCMC convergence diagnostics
(Monte Carlo error, effective sample size, Rhat).
}
