% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm_stan.R
\name{rlm_stan}
\alias{rlm_stan}
\title{Bayesian robust linear regression with Stan}
\usage{
rlm_stan(x, y, family = "Student's t", nuDist = c("fixed", "gamma",
  "uniform", "poisson"), ..., fixedNu, shape, rate, scale, nuMax, lambda)
}
\arguments{
\item{x}{Numeric matrix of input values. This matrix should not contain
a column of 1's.}

\item{y}{Numeric vector of output values.}

\item{family}{Character. Right now, only "Student's t is supported.}

\item{nuDist}{Character, either "fixed" for a fixed value of nu, "continuous", in which case
the prior distribution for nu is a gamma distribution, "uniform", in which case the prior
distribution for nu is a discrete uniform, or "poisson", in which case the prior distribution
for nu is a Poisson distribution truncated at 1 (no 0 values for nu).}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian robust linear regression with Stan
}
\examples{
\dontrun{
x <- runif(30, 0, 10)
y <- 1 + 2*x + rt(30, 4)
rlmFitFixed <- rlm_stan(x, y, family = "Student's t", nuDist = "fixed", fixedNu = 4)
rlmFitCont <- rlm_stan(x, y, family = "Student's t", nuDist = "gamma", shape = 2, rate = .5)
rlmFitDisc <- rlm_stan(x, y, family = "Student's t", nuDist = "uniform", nuMax = 30)
rlmFitDiscP <- rlm_stan(x, y, family = "Student's t", nuDist = "poisson", lambda = 4)
}
}
