% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmfit-class.R
\name{summary.lmfit}
\alias{summary.lmfit}
\title{Summary method for lmfit objects}
\usage{
\method{summary}{lmfit}(object, pars, probs = c(0.025, 0.25, 0.5, 0.75,
  0.975), use_cache = TRUE, ...)
}
\description{
Summaries of parameter estimates and MCMC convergence diagnostics
(Monte Carlo error, effective sample size, Rhat).
}
