% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlmfit-class.R
\name{summary.rlmfit}
\alias{summary.rlmfit}
\title{Summary method for rlmfit objects}
\usage{
\method{summary}{rlmfit}(object, pars, probs = c(0.025, 0.25, 0.5, 0.75,
  0.975), use_cache = TRUE, ...)
}
\description{
Summaries of parameter estimates and MCMC convergence diagnostics
(Monte Carlo error, effective sample size, Rhat).
}
