% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm_stan.R
\name{rlm_stan}
\alias{rlm_stan}
\title{Bayesian robust linear regression with Stan}
\usage{
rlm_stan(x, y, nu = c("fixed", "continuous", "discrete"), ...)
}
\arguments{
\item{x}{Numeric matrix of input values. This matrix should not contain
a column of 1's.}

\item{y}{Numeric vector of output values.}

\item{nu}{Character, either "fixed" for a fixed value of nu (I'm going to make it 4,
since this is a fake package), "continuous", in which case the prior distribution
for nu is gamma(shape = 2, rate = 0.5), for an expected value of 4, or "discrete", in
which case the prior distribution for nu is a discrete uniform(1, 10). In a real package,
I'd give more input options (fixed parameter value, shape, rate, upper bound, ...),
but I don't care here}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian robust linear regression with Stan
}
\examples{
\dontrun{
rlmFit <- rlm_stan(x = runif(30, 0, 10), y = 1 + 2*x + rt(30, 4))
}
}
